refactor: separate hook for isSpeaking feature
